create lib/rclex/nifs.ex and copy NIF funcs to it
